+++ src/main/java/csc1035/project2/Booking.java
@@ -1,5 +1,9 @@
 package csc1035.project2;
 
+import net.bytebuddy.implementation.bind.annotation.Default;
+import org.hibernate.annotations.ColumnDefault;
+import org.hibernate.annotations.GenericGenerator;
+
 import javax.persistence.*;
 import java.sql.Timestamp;
 
@@ -18,7 +22,9 @@ import java.sql.Timestamp;
 @Table(name = "Booking")
 public class Booking {
     @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @ColumnDefault(value = "a")
+    @GeneratedValue(generator="system-uuid")
+    @GenericGenerator(name="system-uuid", strategy = "uuid")
     @Column(name = "Booking_ID", nullable = false)
     private String id;
 
@@ -32,7 +38,7 @@ public class Booking {
     private Timestamp end;
 
     @ManyToOne
-    @JoinColumn(nullable = false)
+    @JoinColumn(nullable = false, insertable = false, updatable = false)
     private Room room;
 
     @OneToOne
@@ -112,5 +118,13 @@ public class Booking {
 
     public void setEnd(Timestamp end) {
         this.end = end;
+    }
+
+    public Room getRoom() {
+        return room;
+    }
+
+    public void setRoom(Room room) {
+        this.room = room;
     }
 }