+++ src/main/java/csc1035/project2/ModuleRequirements.java
@@ -1,6 +1,7 @@
 package csc1035.project2;
 
 import javax.persistence.*;
+import java.util.Date;
 
 /**
  * A persistent class made to handle the Module_Requirements table.
@@ -33,6 +34,9 @@ public class ModuleRequirements {
     @Column(name = "Practical_Length",nullable = false)
     private int plength;
 
+    @Column(name = "Commencing", nullable = false)
+    private Date commencing;
+
     @OneToOne(mappedBy = "moduleRequirements")
     private Modules modules;
 
@@ -46,12 +50,14 @@ public class ModuleRequirements {
      * @param practicals The number of practicals.
      * @param plength The practical length.
      */
-    public ModuleRequirements(String id, int lecturers, int llength, int practicals, int plength) {
+    public ModuleRequirements(String id, int lecturers, int llength,
+                              int practicals, int plength, Date commencing) {
         this.id = id;
         this.lecturers = lecturers;
         this.llength = llength;
         this.practicals = practicals;
         this.plength = plength;
+        this.commencing = commencing;
     }
 
     /**
@@ -114,5 +120,13 @@ public class ModuleRequirements {
 
     public void setPlength(int plength) {
         this.plength = plength;
+    }
+
+    public Date getCommencing() {
+        return commencing;
+    }
+
+    public void setCommencing(Date commencing) {
+        this.commencing = commencing;
     }
 }